LPS-184962 Extract common parts to util directory to reuse in new component
